+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/api/inntekt/rapporterInntekt.ts
@@ -1,12 +1,9 @@
-import {
-    UngdomsytelseController as UngdomsytelseControllerService,
-    UngdomsytelseInntektsrapportering,
-} from '@navikt/k9-brukerdialog-prosessering-api';
+import { ungdomsytelse } from '@navikt/k9-brukerdialog-prosessering-api';
 import { handleApiError } from '@navikt/ung-common';
 
-export const rapporterInntekt = async (data: UngdomsytelseInntektsrapportering): Promise<void> => {
+export const rapporterInntekt = async (data: ungdomsytelse.UngdomsytelseInntektsrapportering): Promise<void> => {
     try {
-        await UngdomsytelseControllerService.inntektrapportering({ body: data, headers: {} as any });
+        await ungdomsytelse.UngdomsytelseController.inntektrapportering({ body: data, headers: {} as any });
         return Promise.resolve();
     } catch (e) {
         throw handleApiError(e, 'rapporterInntekt');